Add basic project structure. Code from previous repo forthcoming.
Three main projects outlined:
Bridge-API - This is what clients will build against. I'm thinking a mostly annotation driven framework. It's a lot of reflection overhead, but we mostly pay that cost once at load time.
Bridge-Core - Where the main logic to interact with the 7-days assemblies and provide the framework will live.
Bridge-Patcher - The patcher that we will provide while we attempt to create better solutions.